+++ xmchord/model/action_reader.h
@@ -66,6 +66,13 @@ class ActionReader {
                                     int offset_start_comment);
 
   std::string WrapOutputLine(const std::string &line);
+
+  void AppendSegmentPreWrapped(const std::string &line,
+                               size_t len_chording_identifier,
+                               const std::string &segment,
+                               unsigned long segment_len,
+                               std::string &line_wrapped,
+                               uint16_t &current_line_length);
 };
 
 }  // namespace model