+++ CHANGELOG.md
@@ -9,6 +9,7 @@ V0.4.1 - Not released yet
 * Add action: ◢ + ENTER = Click left mouse button three times (select line)
 * Actions listing: improve padding between filename, chording identifier and descriptions
 * Change Claws Mail action: ◢ + H = Toggle "Hide read messages" (was: "hide read threads")
+* Add recognition of held modifier keys (left/right alt, ctrl, shift) 
 
 V0.4.0 - 2021/01/12
 -------------------